feat: open composer modal from Cmd+J with GH URL support
Cmd+J's create-worktree action now opens the NewWorkspaceComposerCard
as a modal instead of navigating to the full-page new-workspace view.
When the query is a GitHub issue/PR URL (or raw number), it jumps to
any existing linked worktree; otherwise it resolves the work item via
gh and prefills the composer. Page and palette share a useComposerState
hook so composer state, effects, and create side-effects live in one
place.
