+++ src/renderer/src/components/WorktreeJumpPalette.tsx
@@ -11,6 +11,9 @@ import {
   CommandItem
 } from '@/components/ui/command'
 import { branchName } from '@/lib/git-utils'
+import { parseGitHubIssueOrPRNumber, parseGitHubIssueOrPRLink } from '@/lib/github-links'
+import { getLinkedWorkItemSuggestedName } from '@/lib/new-workspace'
+import type { LinkedWorkItemSummary } from '@/lib/new-workspace'
 import { sortWorktreesSmart } from '@/components/sidebar/smart-sort'
 import StatusIndicator from '@/components/sidebar/StatusIndicator'
 import { cn } from '@/lib/utils'
@@ -131,7 +134,7 @@ function findBrowserSelection(
 export default function WorktreeJumpPalette(): React.JSX.Element | null {
   const visible = useAppStore((s) => s.activeModal === 'worktree-palette')
   const closeModal = useAppStore((s) => s.closeModal)
-  const openNewWorkspacePage = useAppStore((s) => s.openNewWorkspacePage)
+  const openModal = useAppStore((s) => s.openModal)
   const worktreesByRepo = useAppStore((s) => s.worktreesByRepo)
   const repos = useAppStore((s) => s.repos)
   const tabsByWorktree = useAppStore((s) => s.tabsByWorktree)
@@ -473,13 +476,133 @@ export default function WorktreeJumpPalette(): React.JSX.Element | null {
 
   const handleCreateWorktree = useCallback(() => {
     skipRestoreFocusRef.current = true
-    closeModal()
-    // Why: we open the full-page new-workspace view in a microtask so Radix
-    // fully unmounts before the next autofocus cycle runs, avoiding focus churn.
-    queueMicrotask(() =>
-      openNewWorkspacePage(createWorktreeName ? { prefilledName: createWorktreeName } : {})
-    )
-  }, [closeModal, createWorktreeName, openNewWorkspacePage])
+    const trimmed = createWorktreeName.trim()
+    const ghLink = parseGitHubIssueOrPRLink(trimmed)
+    const ghNumber = parseGitHubIssueOrPRNumber(trimmed)
+
+    const openComposer = (data: Record<string, unknown>): void => {
+      closeModal()
+      // Why: defer opening so Radix fully unmounts the palette's dialog before
+      // the composer modal mounts, avoiding focus churn between the two.
+      queueMicrotask(() => openModal('new-workspace-composer', data))
+    }
+
+    // Case 1: user pasted a GH issue/PR URL.
+    if (ghLink) {
+      const { slug, number } = ghLink
+      const state = useAppStore.getState()
+
+      // Why: the existing-worktree check only needs the issue/PR number, which
+      // is repo-agnostic on the worktree meta side. We don't currently cache a
+      // repo-slug map, so slug-matching against a specific repo happens
+      // implicitly when we pick a repo for the `gh workItem` lookup below.
+      const allWorktrees = Object.values(state.worktreesByRepo).flat()
+      const matches = allWorktrees.filter(
+        (w) => !w.isArchived && (w.linkedIssue === number || w.linkedPR === number)
+      )
+      const activeMatch = matches.find((w) => w.repoId === state.activeRepoId) ?? matches[0]
+      if (activeMatch) {
+        closeModal()
+        activateAndRevealWorktree(activeMatch.id)
+        return
+      }
+
+      // Resolve via gh.workItem: prefer the active repo, else the first eligible.
+      const eligibleRepos = state.repos.filter((r) => isGitRepoKind(r))
+      const repoForLookup =
+        (state.activeRepoId && eligibleRepos.find((r) => r.id === state.activeRepoId)) ||
+        eligibleRepos[0]
+      if (!repoForLookup) {
+        openComposer({ prefilledName: trimmed })
+        return
+      }
+
+      // Why: awaiting inside the user gesture would leave the palette open
+      // indefinitely on slow networks. Close immediately and populate the
+      // composer once the lookup returns.
+      closeModal()
+      void window.api.gh
+        .workItem({ repoPath: repoForLookup.path, number })
+        .then((item) => {
+          const data: Record<string, unknown> = { initialRepoId: repoForLookup.id }
+          if (item) {
+            const linkedWorkItem: LinkedWorkItemSummary = {
+              type: item.type,
+              number: item.number,
+              title: item.title,
+              url: item.url
+            }
+            data.linkedWorkItem = linkedWorkItem
+            data.prefilledName = getLinkedWorkItemSuggestedName(item)
+          } else {
+            // Fallback: we couldn't resolve the URL, just seed the name.
+            data.prefilledName = `${slug.owner}-${slug.repo}-${number}`
+          }
+          queueMicrotask(() => openModal('new-workspace-composer', data))
+        })
+        .catch(() => {
+          queueMicrotask(() =>
+            openModal('new-workspace-composer', { initialRepoId: repoForLookup.id })
+          )
+        })
+      return
+    }
+
+    // Case 2: user typed a raw issue number. Resolve against the active repo.
+    if (ghNumber !== null) {
+      const state = useAppStore.getState()
+      const allWorktrees = Object.values(state.worktreesByRepo).flat()
+      const matches = allWorktrees.filter(
+        (w) => !w.isArchived && (w.linkedIssue === ghNumber || w.linkedPR === ghNumber)
+      )
+      const activeMatch = matches.find((w) => w.repoId === state.activeRepoId) ?? matches[0]
+      if (activeMatch) {
+        closeModal()
+        activateAndRevealWorktree(activeMatch.id)
+        return
+      }
+
+      const repoForLookup =
+        (state.activeRepoId && state.repos.find((r) => r.id === state.activeRepoId)) ||
+        state.repos.find((r) => isGitRepoKind(r))
+      if (!repoForLookup || !isGitRepoKind(repoForLookup)) {
+        openComposer({ prefilledName: trimmed })
+        return
+      }
+
+      closeModal()
+      void window.api.gh
+        .workItem({ repoPath: repoForLookup.path, number: ghNumber })
+        .then((item) => {
+          const data: Record<string, unknown> = { initialRepoId: repoForLookup.id }
+          if (item) {
+            const linkedWorkItem: LinkedWorkItemSummary = {
+              type: item.type,
+              number: item.number,
+              title: item.title,
+              url: item.url
+            }
+            data.linkedWorkItem = linkedWorkItem
+            data.prefilledName = getLinkedWorkItemSuggestedName(item)
+          } else {
+            data.prefilledName = trimmed
+          }
+          queueMicrotask(() => openModal('new-workspace-composer', data))
+        })
+        .catch(() => {
+          queueMicrotask(() =>
+            openModal('new-workspace-composer', {
+              initialRepoId: repoForLookup.id,
+              prefilledName: trimmed
+            })
+          )
+        })
+      return
+    }
+
+    // Case 3: plain name — open composer prefilled.
+    openComposer(trimmed ? { prefilledName: trimmed } : {})
+  }, [closeModal, createWorktreeName, openModal])
 
   const handleCloseAutoFocus = useCallback((e: Event) => {
     e.preventDefault()
